Introduce mongo cluster into main docker-compose file. Implement mongo cluster and mongodb initers
