+++ docker-compose.yml
@@ -69,6 +69,46 @@ x-spark-worker: &base_spark_worker
     - SPARK_LOCAL_STORAGE_ENCRYPTION_ENABLED=no
     - SPARK_SSL_ENABLED=no
 
+x-base-mongodb-cluster-unit: &base-mongodb_cluster-unit
+  <<: *base_service
+  image: mongo
+  expose:
+    - "27017"
+  profiles:
+    - mongo
+    - dev_graduate
+    - prod
+
+x-base-mongodb-cluster-unit-healthcheck: &base-mongodb_cluster-unit-healthcheck
+  healthcheck:
+    test: [ "CMD","mongosh", "--eval", "db.adminCommand('ping')" ]
+    interval: 10s
+    timeout: 10s
+    retries: 5
+
+x-base-mongocfg_dependencies: &mongocfg_dependencies
+  depends_on:
+    mongors1n1:
+      condition: service_healthy
+    mongors1n2:
+      condition: service_healthy
+    mongors1n3:
+      condition: service_healthy
+    mongors2n1:
+      condition: service_healthy
+    mongors2n2:
+      condition: service_healthy
+    mongors2n3:
+      condition: service_healthy
+
+x-base-mongos_dependencies: &mongos_dependencies
+  depends_on:
+    mongocfg1:
+      condition: service_healthy
+    mongocfg2:
+      condition: service_healthy
+    mongocfg3:
+      condition: service_healthy
 
 services:
 
@@ -370,8 +410,8 @@ services:
     
   jupyter-notebook:
     <<: *base_service
-#    build: ./recommender_system/jupyter-notebook/
-    image: jupyter/pyspark-notebook:java-17.0.5
+    build: ./recommender_system/jupyter-notebook/
+#    image: jupyter/pyspark-notebook:java-17.0.5
     container_name: jupyter-notebook
     hostname: jupyter-notebook
     depends_on:
@@ -389,23 +429,159 @@ services:
     ports:
       - ${NOTEBOOK_PORT}:${NOTEBOOK_PORT}
 
-  recommender:
-    <<: *base_service
-    build: ./recommender_system/
-    container_name: recommender
-    restart: unless-stopped
-    profiles:
-      - dev_graduate
+
+  mongors1n1:
+    <<: *base-mongodb_cluster-unit
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongors1n1
+    command: mongod --shardsvr --replSet mongors1 --dbpath /data/db --port 27017
     volumes:
-      - ./recommender_system/log/celery_worker.log:/opt/content/log/celery_worker.log:rw
-      - ./recommender_system/log/celery_beat.log:/opt/content/log/celery_beat.log:rw
-      - ./recommender_system/spark/clickhouse-native-jdbc-shaded-2.6.5.jar:/opt/jars/clickhouse-native-jdbc-shaded-2.6.5.jar
-      - ./recommender_system/spark/clickhouse-jdbc-0.3.2.jar:/opt/jars/clickhouse-jdbc-0.3.2.jar
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/data1:/data/db
+
+  mongors1n2:
+    <<: *base-mongodb_cluster-unit
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongors1n2
+    command: mongod --shardsvr --replSet mongors1 --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/data2:/data/db
+
+  mongors1n3:
+    <<: *base-mongodb_cluster-unit
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongors1n3
+    command: mongod --shardsvr --replSet mongors1 --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/data3:/data/db
+
+  mongors2n1:
+    <<: *base-mongodb_cluster-unit
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongors2n1
+    command: mongod --shardsvr --replSet mongors2 --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/data4:/data/db
+
+  mongors2n2:
+    <<: *base-mongodb_cluster-unit
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongors2n2
+    command: mongod --shardsvr --replSet mongors2 --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/data5:/data/db
+
+  mongors2n3:
+    <<: *base-mongodb_cluster-unit
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongors2n3
+    command: mongod --shardsvr --replSet mongors2 --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/data6:/data/db
+
+  mongocfg1:
+    <<: *base-mongodb_cluster-unit
+    <<: *mongocfg_dependencies
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongocfg1
+    command: mongod --configsvr --replSet mongors1conf --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/config1:/data/db
+
+  mongocfg2:
+    <<: *base-mongodb_cluster-unit
+    <<: *mongocfg_dependencies
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongocfg2
+    command: mongod --configsvr --replSet mongors1conf --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/config2:/data/db
+
+  mongocfg3:
+    <<: *base-mongodb_cluster-unit
+    <<: *mongocfg_dependencies
+    <<: *base-mongodb_cluster-unit-healthcheck
+    container_name: mongocfg3
+    command: mongod --configsvr --replSet mongors1conf --dbpath /data/db --port 27017
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+      - /tmp/mongo_cluster/config3:/data/db
+
+  mongos1:
+    <<: *base-mongodb_cluster-unit
+    <<: *mongos_dependencies
+    container_name: mongos1
+    hostname: mongos1
+    command: mongos --configdb mongors1conf/mongocfg1:27017,mongocfg2:27017,mongocfg3:27017 --port 27017 --bind_ip_all
+    ports:
+      - "27017:27017"
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+
+  mongos2:
+    <<: *base-mongodb_cluster-unit
+    <<: *mongos_dependencies
+    container_name: mongos2
+    hostname: mongos2
+    command: mongos --configdb mongors1conf/mongocfg1:27017,mongocfg2:27017,mongocfg3:27017 --port 27017 --bind_ip_all
+    ports:
+      - "27018:27017"
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+
+  mongo_cluster_initer:
+    <<: *base_service
+    image: mongo
+    container_name: mongo_cluster_initer
+    command: ["/bin/bash", "-c", "chmod +x /opt/init_mongodb_cluster.sh && /opt/init_mongodb_cluster.sh"]
+    volumes:
+      - ./recommender_system/init_mongodb_cluster.sh:/opt/init_mongodb_cluster.sh
+    profiles:
+      - mongo
     depends_on:
-      spark-master:
+      - mongos1
+      - mongos2
+    healthcheck:
+      test: bash -c "[ -f "/opt/ready.flag" ]"
+      interval: 20s
+      timeout: 120s
+      retries: 10
+
+  mongodb_initer:
+    <<: *base_service
+    build:
+      context: ./recommender_system/
+      dockerfile: mongodb_initer/Dockerfile
+    container_name: mongodb_initer
+    profiles:
+      - mongo
+    depends_on:
+      mongo_cluster_initer:
         condition: service_healthy
-      redis:
-        condition: service_healthy
+
+#  recommender:
+#    <<: *base_service
+#    build: ./recommender_system/
+#    container_name: recommender
+#    restart: unless-stopped
+#    profiles:
+#      - dev_graduate
+#    volumes:
+#      - ./recommender_system/log/celery_worker.log:/opt/content/log/celery_worker.log:rw
+#      - ./recommender_system/log/celery_beat.log:/opt/content/log/celery_beat.log:rw
+#      - ./recommender_system/spark/clickhouse-native-jdbc-shaded-2.6.5.jar:/opt/jars/clickhouse-native-jdbc-shaded-2.6.5.jar
+#    depends_on:
+#      spark-master:
+#        condition: service_healthy
+#      redis:
+#        condition: service_healthy
 
 networks:
   graduate:
