+++ recommender_system/mongodb_initer/init_mongodb.sh
@@ -1,0 +1,6 @@
+echo -e ""
+echo -e "\033[97m----------------------------------------\033[00m"
+echo -e "\033[94m[+] Create mongodb collection\033[00m"
+echo -e "$\033[97m----------------------------------------\033[00m"
+
+python src/create_collections.py
