+++ recommender_system/mongodb_initer/requirements.txt
@@ -1,0 +1,5 @@
+python-dotenv==0.20.0
+motor==3.1.1
+backoff==2.2.1
+python-dateutil==2.4
+pydantic==1.10.4
