+++ recommender_system/mongodb_initer/src/core/logger.py
@@ -1,0 +1,7 @@
+import logging
+
+logging.basicConfig(
+    level=logging.INFO, format='%(asctime)s %(levelname)s %(name)s %(message)s', datefmt='%Y-%m-%d %H:%M:%S',
+)
+
+logger = logging.getLogger('MONGODB-INITER')
