+++ recommender_system/mongodb_initer/src/db/commands/create_indexes.json
@@ -1,0 +1,8 @@
+{
+  "db_data.recommendations.createIndexes": {
+    "createIndexes": "recommendations",
+    "indexes": [
+      { "key": { "user_id": 1, "film_id": 1 }, "unique": true, "name": "recommendations film_id_user_id key" }
+    ]
+  }
+}
