--- recommender_system/src/db/receiver/mongodb/commands/create_collections.json
@@ -1,34 +1,0 @@
-{
-  "db_data.recommendations.create":{
-    "create": "recommendations",
-    "clusteredIndex": { "key": { "_id": 1 }, "unique": true, "name": "recommendations clustered key" },
-    "validator": {
-        "$jsonSchema": {
-          "bsonType": "object",
-          "required": [
-            "user_id",
-            "film_id",
-            "score"
-          ],
-          "properties": {
-            "user_id": {
-              "bsonType": "string",
-              "description": "User's UUID"
-            },
-            "film_id": {
-              "bsonType": "string",
-              "description": "Film's id"
-            },
-            "score": {
-              "bsonType": "double",
-              "description": "Score"
-            }
-          }
-        }
-    }
-  },
-  "db_admin.recommendations.shardCollection":{
-    "shardCollection": "recommender.recommendations",
-    "key": { "user_id": "hashed"}
-  }
-}
